genertest-package(genertest)	R Documentation

(Automated) Exam Generation using R

Description

Package that generates tests with questions drawn from a tab-delimited database. It can be used to randomly draw questions, to permute questions and answers, to select questions to include in the tests according to their topic or difficulty, to generate questions containing Sweave syntax.

Details

Package: 	 genertest

Type: 	 Package

Version: 	 1.0

Date: 	 2009-01-14

License: 	 GPL2

LazyLoad: 	 yes

Store your questions in a tab delimited database and use function genertest() to generate tests. Specify topics and difficulty of the questions to be included in the exams, the number of completely different or permuted exams to be generated; many other options are available. The output is a LaTeX or PDF file for each generated test. Function write.all() writes on a file that has the same structure as the exams all the questions included in the database; merge.pdf() can be used to merge PDF files containing different exams in a single file.

Author(s)

Lara Lusa

Maintainer: Lara Lusa <lara.lusa@mf.uni-lj.si>

References

http://sites.google.com/site/lara3107/Home/software/genertest

[Package genertest version 1.0 Index]